fix: clippy errors and add plan execution integration tests
- Fix needless_borrow in interactive.rs:1372,1404 (&agent_busy → agent_busy)
- Fix unnecessary_cast in plan_panel.rs:58 (text_overhead as i32 → text_overhead)
- Add test_plan_act_transition_on_completion: verifies Plan→Act transition
  when all steps complete via successful tool calls
- Add test_plan_step_failure_pauses_execution: verifies failed tool call
  marks step Failed and pauses plan execution

+++ src/cli/tui/plan_panel.rs
@@ -55,7 +55,7 @@ fn build_plan_lines(plan: &PlanState, area: Rect) -> Vec<Line<'static>> {
     let total = plan.steps.len();
     let progress_pct = if total > 0 { (done_count as f64 / total as f64) * 100.0 } else { 0.0 };
     let text_overhead = 25; // "[{}] {:.0}% ({}/{})"
-    let available = area.width as i32 - 2 - text_overhead as i32; // 2 for block padding
+    let available = area.width as i32 - 2 - text_overhead; // 2 for block padding
     let bar_width = if available > 0 { available as usize } else { 20 };
     let filled = (progress_pct / 100.0 * bar_width as f64) as usize;
     let bar_str: String = "█".repeat(filled) + &"░".repeat(bar_width.saturating_sub(filled));
